Created a ConfirmExitState class
This will be used whenever the user either presses the X button or attemps to close the program.
Cancel will bring them back to whatever state they were just on and Exit Program will close the program no matter where they are.

This has been added to State.h as a common button function to be used.