Almost finished the LeaderboardState
The user can now view all users and sort by different heuristics.
They can click on a user to view more information about them but this required tweaks to the Button class mainly it having an id to know which button was pressed.

Also had to fix the bug in the Account where the date/time wasn't being stored correctly.

Only thing left with LeaderboardState is searching for users